+++ admin/tabs/activity/feedback/FeedbackPanel.js
@@ -16,7 +16,7 @@ export const feedbackPanel = hoistCmp.factory(
 
 const modelSpec = {
     persistWith: {localStorageKey: 'xhAdminFeedbackState'},
-    enableColChooser: true,
+    colChooserModel: true,
     enableExport: true,
     emptyText: 'No feedback reported...',
     store: {