+++ admin/tabs/preferences/PreferenceModel.js
@@ -29,7 +29,7 @@ export class PreferenceModel {
     @managed
     gridModel = new RestGridModel({
         persistWith: this.persistWith,
-        enableColChooser: true,
+        colChooserModel: true,
         enableExport: true,
         selModel: 'multiple',
         store: {