+++ desktop/cmp/grid/impl/ColChooserDialog.js
@@ -14,9 +14,10 @@ export const colChooserDialog = hoistCmp.factory({
     model: uses(ColChooserModel),
     className: 'xh-col-chooser-dialog',
 
-    render({model, className, ...props}) {
+    render({model, className}) {
 
-        if (!model.isOpen) return null;
+        const {isOpen, width} = model;
+        if (!isOpen) return null;
 
         return dialog({
             icon: Icon.gridPanel(),
@@ -24,7 +25,8 @@ export const colChooserDialog = hoistCmp.factory({
             isOpen: true,
             onClose: () => model.close(),
             item: colChooser({model}),
-            className
+            className,
+            style: {width}
         });
     }
 });